Field of a record : started.
field is the function that evaluates the accesf to a field of a record.
At the moment, it handles errors while trying accessing to
the field of a "variable".

Example:

    val k := "hello"#x.

Normally, the interpreter does not have to deal with that because this is not
its job. The type checker should gives the warranty that this expression
is not well-typed. In this case, I decided to handle this issue throwing
an exception with the following message for this example:

    flap> val k := "hello"#x.
    val k := "hello"#x.
    Failure("\"hello\"#x is not permitted.")

Author: Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>