hopixInterpreter.ml : Deal with an error case.
When the interpreter evaluates:

    val v := "troll".
    val k := v#x.

field will check if *v* is a reference to an address where
a possible record is stored. If *v* is evaluated as an address, then
this address should be a location of record (not checked yet).
Otherwise, *v* is evaluated as a value/function... and the evaluation is aborted.

So the result of these instruction is :

    flap> val v := "troll".
    val v := "troll".
    (5.00679016113e-06s)
    v = "troll"

    flap> val k := v#x.
    val k := v#x.
    Failure("\"troll\" does not refer to an address")

Like I said in the previous commit, this must not be the work of the interpreter.
The type checker has to handle this issue. But because of the lack of this checker,
I had to handle that manually in the interpreter.

Author : Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>