hopixInterpreter.ml : Evaluate the records that are not in the environment.
The interpreter can handle the record created "on the fly" by the user.
That is to say, it can evaluate:

    val n := { x := 42; y := 'a' }#x.
    val c := { x := 42; y := 'a' }#y.

and return these following results:

n = 42.
c = 'a'.

It possible the evaluation will fail with:

    val z := { x := 42; y := 'a'; b := \j => j + 1 }#b.

because the access of b is evaluated using *j* as a variable in the runtime
environment.

Author : Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>