Redefined the HopixInt32 module.
I created a signature : HopixInt.
HopixInt is a signature, i.e. an interface that defines a set of operations
on interger values. This signature and its module (HopixInt32)
that implements it are only known by HopixInterpreter.

HopixInt32 was just modified to implements HopixInt defining t as Int32.t.
It is necessary to do that in order to use it in *primitives*.

Why this implementation?

 - Integrates the operations of comparison

The first version of *primitives* provided by the teacher just allow
the linkage of the arithmetical operations with the operation provided by Int32.
This version cannot be extensible without changing the interpreter
(and introducing bugs).
So, in order to integrates the operations of comparison
between 32-bit integers (Int32), I defined HopixInt32, that integrates
all of the Int32 arithmetical operations + the comparison operations (<, ≤, =, ≥, >).

 - Extensibility

The interpreter just manipulates integer as Int32, as specified in the specification
document of Hopix. But what does it happend if in a new version of Hopix,
a 16-bit or 64-bit integer must be handled ? It would let the interpreter
be modified in depth.
So, HopixInt was defined to allow different implementations
that handle different kinds of integer.

Author: Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>