Fixed another bullshit priority.
These following operators (<,≤,=,≥,>) are not associative and
all of them have a HIGHER priority than BOOLAND (&&) and BOOLOR (||).

"A declaration of one of the following forms:

    %nonassoc uid1... uidn
    %left uid1... uidn
    %right uid1... uidn

attributes both a priority level and an associativity status to the symbols
uid₁, ..., uidn. The priority level assigned to uid₁, ..., uidn is not defined
explicitly; instead, it is defined to be higher than the priority level
assigned by the previous %nonassoc, %left, or %right declaration,
and lower than that assigned by the next %nonassoc, %left, or %right declaration."
                                                          - Menhir documentation

--