hopix lexer : fixed a possible bug.
These following expression:

    val a := "abc\".
    val b := "\".

must not be accepted at lexing. So, whenthe lexer meets '\' followed by '"',
it put the assertion to false.

Why ?

Getting a string with \" at the end could lead to a side-effect by escaping
the '"' (double quote) and get it into the string.
That is to say, the following expression:

    val a := "abc\".

can be understood by the interpreter as

    abc"

instead of

   abc

--