Hopix Interpreter : Mutually recursive functions.
The first binding of each functions was done at each call of
new_env. So the value of some functions was VUnit.
Because of that, the use of recursive functions was not possible.

I fixed that putting the binding out of new_env in order to get it
be called once.

Consequently, it is possible to get these following instructions :

    rec f := \n => if n = 0 then 1 else n * g (n - 1) fi and g := \x => f x.
    val x := f 5. (* 120 *)
    val x := g 5. (* 120 *)

It is just the factorial function that uses f as the main function
and g as a caller of f.

--