hopixParser.ml : No more conflict.
The conflict was generated because of a bad precedence level and
the separated_nonemty_list rule provided by menhir.
Indeed, the token '?' (QMARK, or question mark) had a too higher priority.
'?' must have the lower priority to be correclty parsed.
Neverhtheless it was not efficient because menhir could not generate the parser.
Thanks to this following forum (thank you Aslam):

https://groups.google.com/forum/?hl=fr#!topic/compilix/89BLY9qC_Ao

I find the solution to fix that.
I had to define my own separated non-empty list, and it works.
No conflict.